feat(common)!: cambiar tmdbId a Integer y actualizar config de seguridad

+++ common/src/main/java/com/watchnext/common/model/ContentRef.java
@@ -19,7 +19,7 @@ public class ContentRef {
 
     // ID para el contenido (Pelicula o Serie)
     @Column(name = "tmdb_id", nullable = false)
-    private Long tmdbId;
+    private Integer tmdbId;
 
     // Enum del tipo de contenido
     @Enumerated(EnumType.STRING)
